Move disk related APIs into utils_disk
This commit moves disk specific apis to utils_disk.
